Fix compilation errors

+++ source/netmessages.cpp
@@ -227,7 +227,7 @@ namespace NetMessages
 		{
 			ClassName *msg = CheckNetmessageType<ClassName>( LUA );
 
-			size_t len = 0;
+			unsigned int len = 0;
 			const char *data = LUA->GetString( 2, &len );
 			std::memcpy(
 				reinterpret_cast<void *>( msg->*M ), data, MaximumLength < len ? MaximumLength : len
